requirements.txt
streamlit
pandas
numpy
pdfplumber
pytesseract
pdf2image
PyMuPDF
fuzzywuzzy
Pillow
layoutparser
langchain
langchain-community
langchain-core
langchain-experimental
tqdm
faiss-cpu
python-docx
python-dotenv
openpyxl
scikit-learn
nltk
ollama